Add files via upload
Part b is in here too. In this third and final part of the MultiQuiz project, we will update our app to display summary counts at the end of the quiz. Once the user reviews the summary, the Back button (or gesture) can be used to return to the quiz again. While viewing the summary activity, the user may optionally reset the quiz before clicking Back to completely start everything again. We will also incorporate several test cases into our project.